+++ src/jsMain/kotlin/com/wadii/screens/providerDetail/ProviderDetailEvent.kt
@@ -1,6 +1,6 @@
 package com.wadii.screens.providerDetail
 
 sealed class ProviderDetailEvent {
-    data class Load(val providerId: Int) : ProviderDetailEvent()
-    data class ToggleFollow(val providerId: Int) : ProviderDetailEvent()
+    data class Load(val providerId: Long) : ProviderDetailEvent()
+    data class ToggleFollow(val providerId: Long) : ProviderDetailEvent()
 }
